// apps/anuglar/src/app/pages/base/config/enums/route-names.ts
export const enum eBaseRouteNames {
  IdentityManagement = 'Base::Menu:IdentityManagement',

  Employees = 'AdministrationService::Menu:Employees',
  EmployeesNew = 'AdministrationService::Menu:EmployeeNew',
  EmployeeList = 'AdministrationService::Menu:EmployeeList',

  Customers = 'AdministrationService::Menu:Customers',
  CustomerNew = 'AdministrationService::Menu:CustomerNew',
  CustomerList = 'AdministrationService::Menu:CustomerList',

  Vendor = 'AdministrationService::Menu:Vendor',
  VendorNew = 'AdministrationService::Menu:VendorNew',
  VendorList = 'AdministrationService::Menu:VendorList',

  Vehicle = 'AdministrationService::Menu:Vehicle',
  VehicleNew = 'AdministrationService::Menu:VehicleNew',
  VehicleList = 'AdministrationService::Menu:VehicleList',

  Company = 'AdministrationService::Menu:Company',

  Address = 'AdministrationService::Menu:Address',

  DocNumbers = 'AdministrationService::Menu:DocNumbers',
}